chore: 공통 axios 모듈 호출 방식으로 디렉토리 변경
