+++ src/features/foods/service/index.ts
@@ -1,29 +1,5 @@
-import axios from 'axios'
-import nprogress from 'nprogress'
+import axios from '@/service/axios'
 import type { Food, FoodDetail } from './food'
-
-nprogress.configure({ showSpinner: false, speed: 500 })
-
-axios.defaults.onDownloadProgress = (e: ProgressEvent) => {
-  const percentage = Math.floor(e.loaded * 1.0) / e.total
-  nprogress.set(percentage)
-}
-
-axios.defaults.onUploadProgress = (e: ProgressEvent) => {
-  const percentage = Math.floor(e.loaded * 1.0) / e.total
-  nprogress.set(percentage)
-}
-
-axios.interceptors.response.use(
-  (response) => {
-    nprogress.done()
-    return response
-  },
-  (error) => {
-    nprogress.done()
-    return Promise.reject(error)
-  },
-)
 
 const foodAPI = axios.create({
   baseURL: 'http://localhost:8080/',
